requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
openpyxl==3.1.5
python-multipart==0.0.9
email-validator==2.2.0
